Moving transition ids into RoutePresenter.
