+++ app/src/main/java/mozilla/lockbox/presenter/AppRoutePresenter.kt
@@ -87,19 +87,8 @@ class AppRoutePresenter(
     }
 
     private fun navigateTo(@IdRes destinationId: Int, args: Bundle? = null) {
-        val srcId = getSourceId(navController, destinationId) ?: return
-
-        if (srcId == destinationId && args == null) {
-            // No point in navigating if nothing has changed.
-            return
-        }
-
-        val transition = findTransitionId(srcId, destinationId) ?: destinationId
-
-        super.navigateToFragment(navController, destinationId, transition, args)
+        super.navigateToFragment(navController, destinationId, args)
     }
-
-
 
     private fun bundle(action: AppWebPageAction): Bundle {
         return AppWebPageFragmentArgs.Builder()
@@ -171,53 +160,6 @@ class AppRoutePresenter(
         }
     }
 
-    private fun findTransitionId(@IdRes from: Int, @IdRes to: Int): Int? {
-        // This maps two nodes in the graph_main.xml to the edge between them.
-        // If a RouteAction is called from a place the graph doesn't know about then
-        // the app will log.error.
-        return when (Pair(from, to)) {
-            Pair(R.id.fragment_welcome, R.id.fragment_item_list) -> R.id.action_to_itemList
-            Pair(R.id.fragment_welcome, R.id.fragment_item_list) -> R.id.action_to_webview
-            Pair(R.id.fragment_welcome, R.id.fragment_locked) -> R.id.action_to_locked
-
-            Pair(R.id.fragment_fxa_login, R.id.fragment_item_list) -> R.id.action_fxaLogin_to_itemList
-            Pair(R.id.fragment_fxa_login, R.id.fragment_fingerprint_onboarding) ->
-                R.id.action_fxaLogin_to_fingerprint_onboarding
-            Pair(R.id.fragment_fxa_login, R.id.fragment_onboarding_confirmation) ->
-                R.id.action_fxaLogin_to_onboarding_confirmation
-
-            Pair(R.id.fragment_fingerprint_onboarding, R.id.fragment_onboarding_confirmation) ->
-                R.id.action_fingerprint_onboarding_to_confirmation
-            Pair(R.id.fragment_fingerprint_onboarding, R.id.fragment_autofill_onboarding) ->
-                R.id.action_onboarding_fingerprint_to_autofill
-
-            Pair(R.id.fragment_autofill_onboarding, R.id.fragment_item_list) -> R.id.action_to_itemList
-
-            Pair(R.id.fragment_onboarding_confirmation, R.id.fragment_item_list) -> R.id.action_to_itemList
-            Pair(R.id.fragment_onboarding_confirmation, R.id.fragment_webview) -> R.id.action_to_webview
-
-            Pair(R.id.fragment_locked, R.id.fragment_item_list) -> R.id.action_to_itemList
-            Pair(R.id.fragment_locked, R.id.fragment_welcome) -> R.id.action_locked_to_welcome
-
-            Pair(R.id.fragment_item_list, R.id.fragment_item_detail) -> R.id.action_itemList_to_itemDetail
-            Pair(R.id.fragment_item_list, R.id.fragment_setting) -> R.id.action_itemList_to_setting
-            Pair(R.id.fragment_item_list, R.id.fragment_account_setting) -> R.id.action_itemList_to_accountSetting
-            Pair(R.id.fragment_item_list, R.id.fragment_locked) -> R.id.action_to_locked
-            Pair(R.id.fragment_item_list, R.id.fragment_filter) -> R.id.action_itemList_to_filter
-            Pair(R.id.fragment_item_list, R.id.fragment_webview) -> R.id.action_to_webview
-
-            Pair(R.id.fragment_item_detail, R.id.fragment_webview) -> R.id.action_to_webview
-
-            Pair(R.id.fragment_setting, R.id.fragment_webview) -> R.id.action_to_webview
-
-            Pair(R.id.fragment_account_setting, R.id.fragment_welcome) -> R.id.action_to_welcome
-
-            Pair(R.id.fragment_filter, R.id.fragment_item_detail) -> R.id.action_filter_to_itemDetail
-
-            else -> null
-        }
-    }
-
     private fun openWebsite(url: String) {
         val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(url))
         activity.startActivity(browserIntent, null)
@@ -228,7 +170,4 @@ class AppRoutePresenter(
         settingIntent.data = settingAction.setting.data
         activity.startActivity(settingIntent, null)
     }
-
-
-
 }
